% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{census}
\alias{census}
\title{County demographics based on 1980 US Census}
\format{A data frame with 17 variables
\itemize{
\item name
}}
\usage{
census
}
\description{
data frame based on the 1980 US Census
}
\keyword{datasets}

