\docType{function}
\name{determineDepth}
\alias{determineDepth}
\title{Determine depth of letter values needed for n observations.}
\arguments{
  \item{n}{number of observation to be shown in the LV
  boxplot}
}
\description{
  Determine depth of letter values needed for n
  observations.
}
\details{
  Supply one of \code{k}, \code{alpha} or \code{perc}.
}

