\name{drawLVplot}
\alias{drawLVplot}
\title{Draw an LV plot}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Draw a letter value boxplot
}
\usage{drawLVplot(x,y,k,out,qu,horizontal,col,...)}
\arguments{
\item{x}{x positions}
\item{y}{y positions}
\item{k}{number of letter value statistics used}
\item{out}{out: outliers}
\item{qu}{quantiles}
\item{horizontal}{display horizontally (TRUE) or vertically (FALSE)}
\item{col}{vector of colours to use}
\item{...}{}
}

\details{}

\examples{}
\keyword{internal}
