\name{drawLVplot}
\alias{drawLVplot}
\title{Draw an LV plot.}
\usage{
  drawLVplot(x, y, k, out, qu, horizontal, col, ...)
}
\arguments{
  \item{x}{x positions}

  \item{y}{y positions}

  \item{k}{number of letter value statistics used}

  \item{out}{indices of outliers}

  \item{qu}{quantiles}

  \item{horizontal}{display horizontally (TRUE) or
  vertically (FALSE)}

  \item{col}{vector of colours to use}
}
\description{
  Draw an LV plot.
}
\keyword{internal}

