% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-lvplot.r, R/stat-lvplot.r
\docType{data}
\name{geom_lvplot}
\alias{GeomLvplot}
\alias{StatLvplot}
\alias{geom_lvplot}
\alias{stat_lvplot}
\title{Side-by-side LV boxplots.}
\format{An object of class \code{GeomLvplot} (inherits from \code{Geom}, \code{ggproto}) of length 6.}
\usage{
geom_lvplot(mapping = NULL, data = NULL, stat = "lvplot",
  position = "dodge", outlier.colour = "black", outlier.shape = 19,
  outlier.size = 1.5, outlier.stroke = 0.5, na.rm = TRUE,
  varwidth = FALSE, show.legend = NA, inherit.aes = TRUE, ...)

GeomLvplot

stat_lvplot(mapping = NULL, data = NULL, geom = "lvplot",
  position = "dodge", na.rm = TRUE, conf = 0.95, percent = NULL,
  k = NULL, show.legend = NA, inherit.aes = TRUE, ...)

StatLvplot
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), is combined with the default mapping at the top level of the
plot. You only need to supply \code{mapping} if there isn't a mapping
defined for the plot.}

\item{data}{A data frame. If specified, overrides the default data frame
defined at the top level of the plot.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{outlier.colour}{Override aesthetics used for the outliers. Defaults
come from \code{geom_point()}.}

\item{outlier.shape}{Override aesthetics used for the outliers. Defaults
come from \code{geom_point()}.}

\item{outlier.size}{Override aesthetics used for the outliers. Defaults
come from \code{geom_point()}.}

\item{outlier.stroke}{Override aesthetics used for the outliers. Defaults
come from \code{geom_point()}.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{varwidth}{if \code{FALSE} (default) draw boxes that are the same size for each group. If
\code{TRUE}, boxes are drawn with widths proportional to the
square-roots of the number of observations in the groups (possibly
weighted, using the \code{weight} aesthetic).}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. There are
  three types of arguments you can use here:

  \itemize{
  \item Aesthetics: to set an aesthetic to a fixed value, like
     \code{color = "red"} or \code{size = 3}.
  \item Other arguments to the layer, for example you override the
    default \code{stat} associated with the layer.
  \item Other arguments passed on to the stat.
  }}

\item{geom, stat}{Use to override the default connection between
\code{geom_lvplot} and \code{stat_lvplot}.}

\item{conf}{confidence level}

\item{percent}{numeric value: percent of data in outliers}

\item{k}{number of letter values shown}
}
\description{
An extension of standard boxplots which draws k letter statistics.
Conventional boxplots (Tukey 1977) are useful displays for conveying rough
information about the central 50\% of the data and the extent of the data.
For moderate-sized data sets (\eqn{n < 1000}), detailed estimates of tail
behavior beyond the quartiles may not be trustworthy, so the information
provided by boxplots is appropriately somewhat vague beyond the quartiles,
and the expected number of ``outliers'' and ``far-out'' values for a
Gaussian sample of size \eqn{n} is often less than 10 (Hoaglin, Iglewicz,
and Tukey 1986). Large data sets (\eqn{n \approx 10,000-100,000}) afford
more precise estimates of quantiles in the tails beyond the quartiles and
also can be expected to present a large number of ``outliers'' (about
\eqn{0.4 + 0.007 n}).
The letter-value box plot addresses both these shortcomings: it conveys
more detailed information in the tails using letter values, only out to the
depths where the letter values are reliable estimates of their
corresponding quantiles (corresponding to tail areas of roughly
\eqn{2^{-i}}); ``outliers'' are defined as a function of the most extreme
letter value shown. All aspects shown on the letter-value boxplot are
actual observations, thus remaining faithful to the principles that
governed Tukey's original boxplot.
}
\section{Computed/reported variables}{

\describe{
  \item{k}{Number of Letter Values used for the display}
  \item{LV}{Name of the Letter Value}
  \item{width}{width of the interquartile box}
}
}
\examples{
library(ggplot2)
p <- ggplot(mpg, aes(class, hwy))
p + geom_lvplot(aes(fill=..LV..)) + scale_fill_brewer()
p + geom_lvplot() + geom_jitter(width = 0.2)
p + geom_lvplot(alpha=1, aes(fill=..LV..)) + scale_fill_brewer()

# Outliers
p + geom_lvplot(varwidth = TRUE, aes(fill=..LV..)) + scale_fill_brewer()
p + geom_lvplot(fill = "grey80", colour = "black")
p + geom_lvplot(outlier.colour = "red", outlier.shape = 1)

# Plots are automatically dodged when any aesthetic is a factor
p + geom_lvplot(aes(fill = drv))

\dontrun{
# just for now: read On_Time data into object ot from lvplot paper
library(RColorBrewer)
cols <- c("white",brewer.pal(9, "Greys"), "Red", "Pink")
reds <- brewer.pal(5, "Reds")[-1]
cols <- c("grey50", cols[2:3], reds[1], cols[4:6], reds[2], cols[7:9], reds[3], cols[10:11])

ggplot(data=ot) + geom_lvplot(aes(x=UniqueCarrier, y=sqrt(TaxiOut+TaxiIn),
  fill=..LV..), alpha=1) +
  scale_fill_manual(values=cols)
ggplot(data=ot) + geom_lvplot(aes(x=factor(DayOfWeek), y=sqrt(TaxiOut+TaxiIn),
  fill=..LV..), alpha=1) +
  scale_fill_manual(values=cols) + facet_wrap(~UniqueCarrier)
}
}
\references{
McGill, R., Tukey, J. W. and Larsen, W. A. (1978) Variations of
    box plots. The American Statistician 32, 12-16.
}
\seealso{
\code{\link{stat_quantile}} to view quantiles conditioned on a
  continuous variable.
}
\keyword{datasets}

