\name{LVboxplot}
\alias{LVboxplot}
\title{LV box plot}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
An extension of standard boxplots which draws k letter statistics
}
\usage{LVboxplot(x, ...)}
\arguments{
\item{x}{}
\item{...}{}
}

\details{This is a generic method so please see specific methods for details.}
\seealso{\code{\link{LVboxplot.formula}}, \code{\link{LVboxplot.numeric}}}
\examples{}
\keyword{internal}
