\docType{function}
\name{LVboxplot.numeric}
\alias{LVboxplot.numeric}
\title{Draw a single LV boxplot.}
\arguments{
  \item{x}{numeric vector of data}
}
\description{
  Draw a single LV boxplot.
}
\seealso{
  Other letter.value.boxplots: \code{\link{LVboxplot}},
  \code{\link{LVboxplot.formula}}
}

