\name{LVboxplot}
\alias{LVboxplot}
\title{LV box plot.}
\usage{
  LVboxplot(x, ...)
}
\description{
  An extension of standard boxplots which draws k letter
  statistics.
}
\details{
  This is a generic method: please see specific methods for
  details.
}
\seealso{
  Other letter.value.boxplots:
  \code{\link{LVboxplot.formula}},
  \code{\link{LVboxplot.numeric}}
}
\keyword{internal}

