% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvplot-package.R
\docType{package}
\name{lvplot-package}
\alias{lvplot}
\alias{lvplot-package}
\title{lvplot: Letter Value 'Boxplots'}
\description{
Implements the letter value 'boxplot' which extends the standard 'boxplot' to deal with both larger and smaller number of data points by dynamically selecting the appropriate number of letter values to display.
}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Authors:
\itemize{
  \item Heike Hofmann \email{hofmann@iastate.edu}
}

}
\keyword{internal}
