% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{census}
\alias{census}
\title{County demographics based on 1980 US Census}
\format{A data frame with 10 variables
\describe{
\item{county}{County name}
\item{FIPS}{FIPS county code}
\item{Latitude,Longitude}{Geographic location of county centers}
\item{JanTmp,JulTmp}{(normalized) Temperatures in January & July}
\item{JanSun,JulSun}{(normalized) Sunshine measurment in January & July}
\item{Elevtn}{Elevation above sea level}
\item{totalpop}{Population}
}}
\usage{
census
}
\description{
County level statistics based on the 1980 US Census.
}
\keyword{datasets}

