% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/draw.r
\name{drawLVplot}
\alias{drawLVplot}
\title{Draw an LV plot.}
\usage{
drawLVplot(x, y, k, out, qu, horizontal, col, width = 0.9, median.col, ...)
}
\arguments{
\item{x}{x positions}

\item{y}{y positions}

\item{k}{number of letter value statistics used}

\item{out}{indices of outliers}

\item{qu}{quantiles}

\item{horizontal}{display horizontally (TRUE) or vertically (FALSE)}

\item{col}{vector of colours to use}

\item{width}{height/width of box}
}
\description{
Draw an LV plot.
}
\keyword{internal}

