% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.r
\name{drawLVplot}
\alias{drawLVplot}
\title{Draw an LV plot.}
\usage{
drawLVplot(x, y, k, out, qu, horizontal, col, border = "grey50",
  width = 0.9, width.method = "linear", median.col, ...)
}
\arguments{
\item{x}{x positions}

\item{y}{y positions}

\item{k}{number of letter value statistics used}

\item{out}{indices of outliers}

\item{qu}{quantiles}

\item{horizontal}{display horizontally (TRUE) or vertically (FALSE)}

\item{col}{vector of colours to use}

\item{border}{vector of colours to use}

\item{width}{maximum height/width of box}

\item{width.method}{one of 'linear', 'height' or 'area'. Methods 'height' and 'area' ensure that these dimension are proportional to the number of observations within each box.}
}
\description{
Draw an LV plot.
}
\keyword{internal}

