% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.r
\name{lvtable}
\alias{lvtable}
\title{Compute table of k letter values for vector x}
\usage{
lvtable(x, k, alpha = 0.95)
}
\arguments{
\item{x}{input numeric vector}

\item{k}{number of letter values to compute}

\item{alpha}{alpha-threshold for confidence level}
}
\description{
Compute table of k letter values for vector x
}
