\docType{function}
\name{outputLVplot}
\alias{outputLVplot}
\title{Compute letter value summary table.}
\arguments{
  \item{x}{numeric vector}
}
\value{
  \item{letter.val}{letter value statistic, distinguishes
  between upper and lower LV statistic for all statistics
  but the median} \item{conf.int}{confidence interval of
  corresponding letter value statistic} \item{out}{list of
  defined outliers}
}
\description{
  Compute letter value summary table.
}

